[backend]move test file(#4806)
