--- opencti-platform/opencti-graphql/tests/02-integration/05-parser/default-values/hashes.csv
@@ -1,2 +1,0 @@
-sha256,md5,sha-1
-9e403e07a54437e37fe5dc47f10e93370e4b685fda8bf6fe00013a519bd228ce,ba69669818ef9ccec174d647a8021a7b,b8e74921d7923c808a0423e6e46807c4f0699b6e
